+++ lib/src/models/tex_view_rendering_engine.dart
@@ -2,20 +2,52 @@ import 'dart:convert';
 
 class TeXViewRenderingEngine {
   final String configurations;
-  final String _engineName;
+  final String name;
 
   const TeXViewRenderingEngine.katex({this.configurations})
-      : this._engineName = "katex";
+      : this.name = "katex";
 
-  const TeXViewRenderingEngine.mathjax()
-      : this._engineName = "mathjax",
-        configurations = null;
+  const TeXViewRenderingEngine.mathjax(
+      {this.configurations = Configurations.mathjaxDefaultConfiguration})
+      : this.name = "mathjax";
 
   String getConfigurations() {
     return json.encode(configurations ?? "{}");
   }
 
-  String getEngineName() {
-    return this._engineName;
+  static String dd = "dfd";
+}
+
+class Configurations {
+  static const String mathjaxDefaultConfiguration = """
+  {
+      showMathMenu: false,
+      delayStartupUntil: "configured",
+      tex2jax: {
+          ignoreClass: "tex2jax_ignore",
+      },
+      jax: ["input/TeX", "output/SVG"],
+      extensions: ["tex2jax.js", "MathEvents.js"],
+      TeX: {
+          extensions: ["mhchem.js", "noErrors.js", "noUndefined.js", "autoload-all.js"],
+          noErrors: {disabled: true}
+      },
+      MathMenu: {
+          showRenderer: false
+      },
+      messageStyle: "none",
+      SVG: {
+          blacker: 0,
+          font: "Gyre-Pagella",
+          linebreaks: {
+              automatic: true,
+              width: "container"
+          },
+          scale: 90,
+          useFontCache: true,
+          useGlobalCache: true,
+          matchFontHeight: true
+      }
   }
+  """;
 }