+++ lib/src/views/tex_view.dart
@@ -7,50 +7,41 @@ import 'package:flutter_tex/src/views/tex_view_mobile.dart'
 
 ///A Flutter Widget to render Mathematics / Maths, Physics and Chemistry, Statistics / Stats Equations based on LaTeX with full HTML and JavaScript support.
 class TeXView extends StatefulWidget {
-  final Key key;
-
   /// A list of TeXViewChild.
   final TeXViewWidget child;
 
   /// Style TeXView Widget with [TeXViewStyle].
   final TeXViewStyle style;
 
+  /// Register fonts.
   final List<TeXViewFont> fonts;
 
   /// Render Engine to render TeX.
   final TeXViewRenderingEngine renderingEngine;
 
-  /// Fixed Height for TeXView. (Avoid using fixed height for TeXView, let it to adopt the height by itself)
-  final double height;
-
   /// Show a loading widget before rendering completes.
-  final Widget loadingWidget;
-
-  /// Show or hide loadingWidget.
-  final bool showLoadingWidget;
+  final Widget Function(BuildContext context) loadingWidgetBuilder;
 
   /// Callback when TEX rendering finishes.
   final Function(double height) onRenderFinished;
 
   /// Callback when TeXView loading finishes.
-  final Function(String message) onPageFinished;
+  final Function(TeXViewController controller) onTeXViewCreated;
 
   /// Keep widget Alive. (True by default).
   final bool keepAlive;
 
-  const TeXView(
-      {this.key,
-      @required this.child,
-      this.fonts,
-      this.style,
-      this.height,
-      this.loadingWidget,
-      this.showLoadingWidget = false,
-      this.keepAlive,
-      this.onRenderFinished,
-      this.onPageFinished,
-      this.renderingEngine = const TeXViewRenderingEngine.katex()})
-      : super(key: key);
+  const TeXView({
+    Key key,
+    @required this.child,
+    this.fonts,
+    this.style,
+    this.loadingWidgetBuilder,
+    this.keepAlive,
+    this.onRenderFinished,
+    this.onTeXViewCreated,
+    this.renderingEngine,
+  }) : super(key: key);
 
   @override
   TeXViewState createState() => TeXViewState();