+++ lib/src/widgets/tex_view_video.dart
@@ -3,16 +3,20 @@ import 'package:flutter_tex/src/models/tex_view_widget_meta.dart';
 import 'package:flutter_tex/src/style/style_utils.dart';
 
 class TeXViewVideo extends TeXViewWidget {
+  final String id;
+
   /// Uri for Image.
   final String url;
 
   final String _type;
 
-  const TeXViewVideo.youtube(this.url) : this._type = 'tex-view-video-youtube';
+  const TeXViewVideo.youtube(this.url, {this.id})
+      : this._type = 'tex-view-video-youtube';
 
   @override
   TeXViewWidgetMeta meta() {
-    return TeXViewWidgetMeta(tag: 'div', classList: _type, node: Node.Leaf);
+    return TeXViewWidgetMeta(
+        id: this.id, tag: 'div', classList: _type, node: Node.Leaf);
   }
 
   @override
@@ -20,10 +24,10 @@ class TeXViewVideo extends TeXViewWidget {
 
   @override
   Map toJson() => {
-        'meta': meta().toJson(),
-        'data': _initData(),
-        'style': "max-width: 100%; max-height: 100%; " + teXViewDefaultStyle,
-      };
+    'meta': meta().toJson(),
+    'data': _initData(),
+    'style': "max-width: 100%; max-height: 100%; " + teXViewDefaultStyle,
+  };
 
   String _initData() {
     return """<iframe width="100%" height="100%" frameborder="0" allowfullscreen